:recycle: refactor: 게시글 작성 관련 컴포넌트 폴더 구조 변경
