+++ src/app/school/market/[marketType]/_components/_PostComponents/GroupPurchase.tsx
